# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Mini Banking Platform API
  description: |
    A simplified banking platform API that handles financial transactions with
    double-entry bookkeeping. The system maintains data integrity through ledger
    validation and provides user-friendly error responses following RFC 7807.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: |
        Creates a new user with initial accounts:
        - 1 USD account with $1000.00 initial balance
        - 1 EUR account with €500.00 initial balance
        
        Initial balances are funded from the system cashbook to maintain
        double-entry bookkeeping integrity.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/validation-error"
                title: "Validation Error"
                status: 400
                detail: "Email field is required"
                instance: "/auth/register"
        '409':
          description: User already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/user-already-exists"
                title: "User Already Exists"
                status: 409
                detail: "User with email user@example.com already exists"
                instance: "/auth/register"
                email: "user@example.com"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate user
      description: Authenticates a user and returns a JWT token.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Invalid credentials
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/invalid-credentials"
                title: "Invalid Credentials"
                status: 401
                detail: "The provided email or password is incorrect"
                instance: "/auth/login"
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user info
      description: Returns information about the currently authenticated user.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/unauthorized"
                title: "Unauthorized"
                status: 401
                detail: "Authentication required"
                instance: "/auth/me"
  /accounts:
    get:
      tags:
        - Accounts
      summary: List user's accounts
      description: Returns all accounts belonging to the authenticated user with their current balances.
      operationId: listAccounts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /accounts/{accountId}/balance:
    get:
      tags:
        - Accounts
      summary: Get account balance
      description: Returns the current balance of the specified account.
      operationId: getAccountBalance
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden - account does not belong to user
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/forbidden"
                title: "Forbidden"
                status: 403
                detail: "You do not have access to this account"
                instance: "/accounts/123e4567-e89b-12d3-a456-426614174000/balance"
        '404':
          description: Account not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/account-not-found"
                title: "Account Not Found"
                status: 404
                detail: "Account 123e4567-e89b-12d3-a456-426614174000 not found"
                instance: "/accounts/123e4567-e89b-12d3-a456-426614174000/balance"
                accountId: "123e4567-e89b-12d3-a456-426614174000"
  /transactions/transfer:
    post:
      tags:
        - Transactions
      summary: Transfer money between users
      description: |
        Transfers money from the authenticated user's account to another user's account.
        Both accounts must have the same currency. The operation is atomic and maintains
        double-entry bookkeeping integrity.
      operationId: transfer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Invalid transfer request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                negativeAmount:
                  summary: Negative amount
                  value:
                    type: "https://minibankingplatform.com/problems/negative-transfer-amount"
                    title: "Negative Transfer Amount"
                    status: 400
                    detail: "Cannot transfer a negative amount: -100.00"
                    instance: "/transactions/transfer"
                    amount: "-100.00"
                currencyMismatch:
                  summary: Currency mismatch
                  value:
                    type: "https://minibankingplatform.com/problems/currency-mismatch"
                    title: "Currency Mismatch"
                    status: 400
                    detail: "USD and EUR are not equal currencies"
                    instance: "/transactions/transfer"
                    sourceCurrency: "USD"
                    targetCurrency: "EUR"
                unsupportedCurrency:
                  summary: Unsupported currency
                  value:
                    type: "https://minibankingplatform.com/problems/unsupported-currency"
                    title: "Unsupported Currency"
                    status: 400
                    detail: "Unsupported currency GBP"
                    instance: "/transactions/transfer"
                    currency: "GBP"
                insufficientFunds:
                  summary: Insufficient funds
                  value:
                    type: "https://minibankingplatform.com/problems/insufficient-funds"
                    title: "Insufficient Funds"
                    status: 400
                    detail: "Account has insufficient funds for this transfer"
                    instance: "/transactions/transfer"
                    available: "500.00"
                    required: "1000.00"
                    currency: "USD"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Account not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/account-not-found"
                title: "Account Not Found"
                status: 404
                detail: "Account 123e4567-e89b-12d3-a456-426614174000 not found"
                instance: "/transactions/transfer"
                accountId: "123e4567-e89b-12d3-a456-426614174000"
  /transactions/exchange:
    post:
      tags:
        - Transactions
      summary: Exchange currency within user's accounts
      description: |
        Exchanges money between the authenticated user's accounts with different currencies.
        Uses a fixed exchange rate: 1 USD = 0.92 EUR. The operation is atomic and maintains
        double-entry bookkeeping integrity through cashbook accounts.
      operationId: exchange
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRequest'
      responses:
        '200':
          description: Exchange completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'
        '400':
          description: Invalid exchange request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                negativeAmount:
                  summary: Negative amount
                  value:
                    type: "https://minibankingplatform.com/problems/negative-exchange-amount"
                    title: "Negative Exchange Amount"
                    status: 400
                    detail: "Cannot exchange a negative amount: -100.00"
                    instance: "/transactions/exchange"
                    amount: "-100.00"
                sameCurrency:
                  summary: Same currency exchange
                  value:
                    type: "https://minibankingplatform.com/problems/same-currency-exchange"
                    title: "Same Currency Exchange"
                    status: 400
                    detail: "Cannot exchange within the same currency: USD"
                    instance: "/transactions/exchange"
                    currency: "USD"
                insufficientFunds:
                  summary: Insufficient funds
                  value:
                    type: "https://minibankingplatform.com/problems/insufficient-funds"
                    title: "Insufficient Funds"
                    status: 400
                    detail: "Account has insufficient funds for this exchange"
                    instance: "/transactions/exchange"
                    available: "500.00"
                    required: "1000.00"
                    currency: "USD"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Account not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /transactions/exchange/calculate:
    get:
      tags:
        - Transactions
      summary: Calculate exchange amount
      description: |
        Calculates how much will be received when exchanging the given amount
        from source currency to target currency using the current exchange rate.
        This is a read-only operation for preview purposes.
      operationId: calculateExchange
      security:
        - BearerAuth: []
      parameters:
        - name: amount
          in: query
          required: true
          description: Amount to exchange
          schema:
            type: string
            pattern: ^\d+(\.\d{1,2})?$
            example: "100.00"
        - name: sourceCurrency
          in: query
          required: true
          description: Source currency
          schema:
            $ref: '#/components/schemas/Currency'
        - name: targetCurrency
          in: query
          required: true
          description: Target currency
          schema:
            $ref: '#/components/schemas/Currency'
      responses:
        '200':
          description: Exchange calculation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeCalculation'
        '400':
          description: Invalid calculation request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                sameCurrency:
                  summary: Same currency
                  value:
                    type: "https://minibankingplatform.com/problems/same-currency-exchange-rate"
                    title: "Same Currency Exchange Rate"
                    status: 400
                    detail: "Exchange rate cannot have the same source and target currency: USD"
                    instance: "/transactions/exchange/calculate"
                    currency: "USD"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /system/reconcile:
    get:
      tags:
        - System
      summary: Reconciliation report
      description: |
        Performs a full reconciliation check of the double-entry bookkeeping system.
        Verifies:
        1. The sum of all ledger entries for each currency equals zero
        2. Each account's balance matches the sum of its ledger entries
        
        This is a read-only operation for auditing and monitoring purposes.
      operationId: reconcile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Reconciliation report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationReport'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoints
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
          x-oapi-codegen-extra-tags:
            validate: "required,email"
        password:
          type: string
          minLength: 8
          example: "securePassword123"
          x-oapi-codegen-extra-tags:
            validate: "required,min=8"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
          x-oapi-codegen-extra-tags:
            validate: "required,email"
        password:
          type: string
          example: "securePassword123"
          x-oapi-codegen-extra-tags:
            validate: "required"
...
    AuthResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        token:
          type: string
          description: JWT token for authentication
...